cleaned data_hubs tabs. made the folder strcuture much more intutive and clean
