+++ pages/data_hub/tabs/manipulation_tabs.py
@@ -1,171 +1,31 @@
-import streamlit as st
-from data_utils import *
-from data_analysis import *
-
 import json
+import streamlit as st
 
-from models import get_db, DQRule, DatasetAction, DatasetVersion  # Import the Dataset model and database session
+import pandas as pd
+
+from models import get_db, DatasetAction  # Import the Dataset model and database session
 from sqlalchemy.orm import Session
 
+from sklearn.preprocessing import StandardScaler, MinMaxScaler, OneHotEncoder, LabelEncoder
+from scipy.stats import zscore
 
-# Try to import necessary packages and install if not available
-# Try to import necessary packages and install if not available
-try:
-    import plotly.express as px
-    from sklearn.preprocessing import StandardScaler, MinMaxScaler, OneHotEncoder, LabelEncoder
-    from scipy.stats import zscore
-except ModuleNotFoundError as e:
-    import subprocess
-    import sys
-    missing_package = str(e).split("'")[1]  # Get the missing package name
-    
-    # Correctly handle the sklearn package by installing scikit-learn
-    if missing_package == 'sklearn':
-        subprocess.check_call([sys.executable, "-m", "pip", "install", "scikit-learn"])
-    elif missing_package == 'plotly':
-        subprocess.check_call([sys.executable, "-m", "pip", "install", "plotly"])
-    elif missing_package == 'scipy':
-        subprocess.check_call([sys.executable, "-m", "pip", "install", "scipy"])
-    
-    # Reimport after installation
-    if missing_package == "plotly":
-        import plotly.express as px
-    elif missing_package == "sklearn":
-        from sklearn.preprocessing import StandardScaler, MinMaxScaler, OneHotEncoder, LabelEncoder
-    elif missing_package == "scipy":
-        from scipy.stats import zscore
-
-# Function to display the summary as tiles
-def display_summary_tiles(summary):
-    """Displays the summary statistics in a tile format."""
-    col1, col2, col3 = st.columns(3)
-    col1.metric("Number of Rows", summary['Number of Rows'])
-    col1.metric("Number of Columns", summary['Number of Columns'])
-    col2.metric("Missing Values", summary['Missing Values'])
-    col2.metric("Duplicate Rows", summary['Duplicate Rows'])
-    col3.metric("Memory Usage (MB)", summary['Memory Usage (MB)'])
-
-# Function to display the column-level summary and distribution side by side
-def display_column_summary(df, column):
-    """Displays the summary of the selected column with distribution plots."""
-    summary = column_summary(df, column)
-    col1, col2 = st.columns(2)
-    
-    with col1:
-        st.write("Data Type:", summary['Data Type'])
-        st.write("Unique Values:", summary['Unique Values'])
-        st.write("Missing Values:", summary['Missing Values'])
-        st.write("Mean:", summary['Mean'])
-        st.write("Median:", summary['Median'])
-        st.write("Mode:", summary['Mode'])
-        st.write("Standard Deviation:", summary['Standard Deviation'])
-        st.write("Min:", summary['Min'])
-        st.write("Max:", summary['Max'])
-    
-    with col2:
-        st.subheader(f"Distribution of {column}")
-        if pd.api.types.is_numeric_dtype(df[column]):
-            fig = px.histogram(df, x=column, marginal="box", nbins=30, title=f'Distribution of {column}')
-        else:
-            fig = px.histogram(df, x=column, color=column, title=f'Distribution of {column}')
-        st.plotly_chart(fig, use_container_width=True)
-
-# Function to handle the first tab (Data Summary)
-def handle_data_summary_tab(filtered_data):
-    """Handles all content and logic within the Data Summary tab."""
-    st.header("Data Summary")
-    
-    # Display summary statistics in tiles
-    summary = generate_summary(filtered_data)
-    display_summary_tiles(summary)
-    
-    # Combined accordion with two tabs for detailed statistics and column-level summary
-    st.header("Detailed Analysis")
-    with st.expander("View Detailed Analysis", expanded=False):
-        sub_tabs = st.tabs(["Detailed Statistics", "Column-Level Summary"])
-        
-        with sub_tabs[0]:
-            st.subheader("Detailed Statistics")
-            st.dataframe(detailed_statistics(filtered_data), use_container_width=True)
-        
-        with sub_tabs[1]:
-            st.subheader("Column-Level Summary")
-            selected_column = st.selectbox("Select Column", filtered_data.columns)
-            if selected_column:
-                display_column_summary(filtered_data, selected_column)
-
-def handle_data_analysis_tab(filtered_data):
-    """Handles all content and logic within the Data Analysis Tab."""
-    st.header("Data Analysis")
-    st.write("This tab will host various analysis tools, such as univariate, bivariate, and multivariate analysis, along with other advanced data analysis features.")
-    
-    # Dropdown for analysis options
-    analysis_option = st.selectbox(
-        "Select an Analysis Type",
-        options=[
-            "Univariate Analysis",
-            "Bivariate Analysis",
-            "Multivariate Analysis",
-            "Correlation Analysis",
-            "Cross Tabulation"
-        ]
+def log_action(version_id, action_type, parameters):
+    """Logs the action to the database."""
+    new_action = DatasetAction(
+        version_id=version_id,
+        action_type=action_type,
+        parameters=json.dumps(parameters)  # Convert parameters to a JSON string
     )
-    
-    # Show the description based on the selected analysis
-    description = {
-        "Univariate Analysis": "Analyze the distribution and summary statistics of individual variables.",
-        "Bivariate Analysis": "Analyze the relationship between two variables.",
-        "Multivariate Analysis": "Analyze relationships involving more than two variables.",
-        "Correlation Analysis": "Analyze correlations between numerical variables.",
-        "Cross Tabulation": "Analyze relationships between categorical variables."
-    }[analysis_option]
-    
-    st.subheader(analysis_option)
-    st.write(f"Description: {description}")
-    st.markdown("---")
-    
-    # Univariate Analysis implementation
-    if analysis_option == "Univariate Analysis":
-        selected_column = st.selectbox("Select Column for Univariate Analysis", filtered_data.columns)
-        if selected_column:
-            display_univariate_analysis(filtered_data, selected_column)
-    
-    # Bivariate Analysis implementation
-    elif analysis_option == "Bivariate Analysis":
-        col1, col2 = st.columns(2)
-        with col1:
-            x_column = st.selectbox("Select X-axis Column", filtered_data.columns)
-        with col2:
-            y_column = st.selectbox("Select Y-axis Column", filtered_data.columns)
-        if x_column and y_column:
-            display_bivariate_analysis(filtered_data, x_column, y_column)
-    # Multivariate Analysis implementation
-    elif analysis_option == "Multivariate Analysis":
-        selected_columns = st.multiselect("Select Columns for Multivariate Analysis", filtered_data.columns)
-        if selected_columns:
-            display_multivariate_analysis(filtered_data, selected_columns)
-    elif analysis_option == "Correlation Analysis":
-        display_correlation_analysis(filtered_data)
-
-# Function to handle the Data Quality tab
-def handle_data_quality_tab(filtered_data, dataset_id):
-    """Handles all content and logic within the Data Quality tab."""
-    st.header("Data Quality Check")
-    
-    # Fetch DQ rules for the selected dataset from the database
     db: Session = next(get_db())
-    rules = db.query(DQRule).filter(DQRule.dataset_id == dataset_id).all()
 
-    # Apply DQ rules and show loader while processing
-    with st.spinner("Applying Data Quality Rules..."):
-        violations = apply_dq_rules(filtered_data, rules)
-    
-    if violations:
-        st.warning("Data Quality Issues Found:")
-        for violation in violations:
-            st.write(f"{violation['severity']}: {violation['message']} in column {violation['column']}")
+    db.add(new_action)
+    db.commit()
+    # After logging the action, update the session state and the history
+    if "actions" in st.session_state:
+        st.session_state.actions.append(new_action)
     else:
-        st.success("No data quality issues found!")
+        st.session_state.actions = [new_action]
+    st.rerun()
 
 # Function to handle the Data Manipulation Tab
 def handle_data_manipulation_tab(filtered_data, selected_version):
@@ -187,26 +47,6 @@ def handle_data_manipulation_tab(filtered_data, selected_version):
             "Replace Values"
         ]
     )
-
-    db: Session = next(get_db())
-
-    def log_action(version_id, action_type, parameters):
-        """Logs the action to the database."""
-        new_action = DatasetAction(
-            version_id=version_id,
-            action_type=action_type,
-            parameters=json.dumps(parameters)  # Convert parameters to a JSON string
-        )
-        db.add(new_action)
-        db.commit()
-        # After logging the action, update the session state and the history
-        if "actions" in st.session_state:
-            st.session_state.actions.append(new_action)
-        else:
-            st.session_state.actions = [new_action]
-        st.rerun()
-        
-
     # Handling each action
     if action == "Rename Column":
         selected_column = st.selectbox("Select Column to Rename", filtered_data.columns)
@@ -303,9 +143,6 @@ def handle_data_manipulation_tab(filtered_data, selected_version):
             log_action(selected_version.id, "Replace Values", {"column": selected_column, "to_replace": to_replace, "replace_with": replace_with})
 
     st.session_state.original_data = filtered_data
-            
-
-
 
 # Function to handle the Preprocessing Tab
 def handle_preprocessing_tab(filtered_data, selected_version):
@@ -322,25 +159,6 @@ def handle_preprocessing_tab(filtered_data, selected_version):
             "Remove Outliers"
         ]
     )
-
-    db: Session = next(get_db())
-
-    def log_action(version_id, action_type, parameters):
-        """Logs the action to the database."""
-        new_action = DatasetAction(
-            version_id=version_id,
-            action_type=action_type,
-            parameters=json.dumps(parameters)  # Convert parameters to a JSON string
-        )
-        db.add(new_action)
-        db.commit()
-        # After logging the action, update the session state and the history
-        if "actions" in st.session_state:
-            st.session_state.actions.append(new_action)
-        else:
-            st.session_state.actions = [new_action]
-        st.rerun()
-
     # Handling each preprocessing action
     if action == "Scale Data":
         selected_columns = st.multiselect("Select Columns to Scale", filtered_data.columns)
