+++ requirements.txt
@@ -1,17 +1,28 @@
 altair==5.4.0
+annotated-types==0.7.0
+anyio==4.4.0
 attrs==24.2.0
 blinker==1.8.2
 cachetools==5.4.0
 certifi==2024.7.4
+chardet==5.2.0
 charset-normalizer==3.3.2
 click==8.1.7
 contourpy==1.2.1
 cycler==0.12.1
+distro==1.9.0
 fonttools==4.53.1
 gitdb==4.0.11
 GitPython==3.1.43
+greenlet==3.0.3
+h11==0.14.0
+h2o==3.46.0.4
+httpcore==1.0.5
+httpx==0.27.0
 idna==3.7
 Jinja2==3.1.4
+jiter==0.5.0
+joblib==1.4.2
 jsonschema==4.23.0
 jsonschema-specifications==2023.12.1
 kiwisolver==1.4.5
@@ -21,27 +32,39 @@ matplotlib==3.9.2
 mdurl==0.1.2
 narwhals==1.4.0
 numpy==2.0.1
+openai==1.41.0
 packaging==24.1
 pandas==2.2.2
 pillow==10.4.0
+plotly==5.23.0
 protobuf==5.27.3
 pyarrow==17.0.0
+pydantic==2.8.2
+pydantic_core==2.20.1
 pydeck==0.9.1
 Pygments==2.18.0
 pyparsing==3.1.2
 python-dateutil==2.9.0.post0
 pytz==2024.1
 referencing==0.35.1
+reportlab==4.2.2
 requests==2.32.3
 rich==13.7.1
 rpds-py==0.20.0
+scikit-learn==1.5.1
+scipy==1.14.0
 seaborn==0.13.2
 six==1.16.0
 smmap==5.0.1
+sniffio==1.3.1
+SQLAlchemy==2.0.32
 streamlit==1.37.1
+tabulate==0.9.0
 tenacity==8.5.0
+threadpoolctl==3.5.0
 toml==0.10.2
 tornado==6.4.1
+tqdm==4.66.5
 typing_extensions==4.12.2
 tzdata==2024.1
 urllib3==2.2.2
